Revert "Remove explicit dex_pc from RecordPcInfo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/be1d5b92-5f24-45f5-a88a-988b2a0213ab). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie114d3b56c7cb713b07a7aa63e2228ffe6dbdc44
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com