Revert "Add option to write zeros instead of sparse."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a48e2170-afff-4a00-9b37-93ea2de788ba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If5ca3f35132a89c32a119ecdba308521e224d99f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com