Revert "Add a one-pass baseline compiler for arm64."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/553df536-74eb-4845-b21c-c250214db686). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4897daba6ba9468671f4ddf003d7c5bb8502b5f2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com