Do most of `ClassLinker::LoadClass()` in native state.
Load class data to temporary storage without holding the
mutator lock to avoid holding up any checkpoints. Then
copy the data to the actual `Class` structures.

Note: The `ArtMethod::{hotness_count_,imt_index_}` was
previously documented as unused and zero-initialized for
abstract class (non-interface) methods in
    https://android-review.googlesource.com/3477431 .
This unused value was actually initialized to the initial
hotness threshold but we're making it zero-initialized now.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 329196666
Change-Id: I07a040944e9137481897b5d9313d56e623a3894f