Revert "Add option to write zeros instead of sparse."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8f074f82-17da-4546-ae02-a2e29c3253f3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6d058f345342429deb5ff4b789eb9f5e448eaca0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com