Revert "Add option to write zeros instead of sparse."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7cf6f2bf-f7ea-423f-b83d-1906130b4590). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4f3171ad70af2ef3cbb10089260074b8712863af
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com