Revert "Add option to write zeros instead of sparse."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b8089a88-981c-4389-a6c2-9a52bb60b22c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e02e6d5e0a255a42b9244023749bdfc09c1ccc8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com