Fix libcore failures with generational CMC
When the heap is found to be full, we skip performing compaction.
The old-gen and mid-gen markers in that case were not being correctly
adjusted. In particular, we had a possibility wherein a young object
is first promoted to old-gen and then in next GC moved back to young-gen.
This is problematic as then we may miss dirtying the card corresponding
to its holder object.

Test: ART_USE_READ_BARRIER=false art/tools/run-libcore-tests.sh --mode=host --variant=X64 --debug --gcstress
Bug: 392794369
Change-Id: I4472b34af3cbf409b2e0478e337c62ffc45c89f5