Use a more concise binary format when dumping long running methods
We used to use a human readable format to dump long running methods.
To make the dumps more concise we are moving to a binary format that
cuts the size by a 3rd. A trace that was about 60K earlier is now about
20K. The format is very similar to the method trace that uses leb
encoding and encodes the incremental difference from the previous entry
both for method ids and the timestamps.

Bug: 352518093
Test: Manual testing with feature enabled and dumping to file and also
to an ANR report.

Change-Id: Ief407dc7dc6f335b0d5add2a1e0f439ad01e318e