Revert "Optimize FindReferenceInfoOf"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f8f535b6-d597-4ae3-9e93-b103c59dcef9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I831f562da9348e34658486b125e81e56d3726c02
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com