Add //apex_available:platform to art tests
Soong was accidentally allowing a corner case where modules that
were not supposed to be available to the platform (due to
apex_available being set to an apex and not including
"//apex_available:platform") were being made available to the
platform as long as no apex modules depended on them.  Add
"//apex_available:platform" explicitly to art tests whose install
rules are hidden when the corner case is fixed.

Bug: 394664568
Test: compare out/soong/installs-${TARGET_PRODUCT}.mk
Change-Id: If0646aa317b1033f5a7c0622fa0fbf89286b4a83