Revert "Fix memory leak when deallocating OatWriter."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9ef3d2e2-1985-4bac-afe6-6531090ffc70). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395919382
Change-Id: Ieb7865ee2409938ef0b70a63b07c86db1d280d47
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com