Revert "Log info about the caller and callee in hiddenapi denial..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1205f02-900c-4c6c-9d81-51f866804348). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: Ie3e4eb41091e6380a9f6bcd63949e2822eeca84e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com