Revert "Allow ART internal libs to load libs in NATIVELOADER_DEF..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1205f02-900c-4c6c-9d81-51f866804348). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: I79e9085594545faf935ed1f2020df08d9f0ed1bc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com