Revert "Fix the loading of libarttest(d).so in 150-loadlibrary a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1205f02-900c-4c6c-9d81-51f866804348). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: I892402d4e8b244124726aaab45c2adffbda18980
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com