Revert "Use libarttest(d)_external to fix 656-annotation-lookup-..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb2b0a1e-355a-4993-9d6b-8701b740e438). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: Ib43fdedf43292ebc1f196f6e52f6154d6f15d1f9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com