Revert "Allow ART internal libs to load libs in NATIVELOADER_DEF..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb2b0a1e-355a-4993-9d6b-8701b740e438). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: Ide4afbf20555eb133f1fc72d183d723da6400e50
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com