Revert "Fix the loading of libarttest(d).so in 150-loadlibrary a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb2b0a1e-355a-4993-9d6b-8701b740e438). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: Ia9d833f83abb71ccba6dfc820c73c566d19434e9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com