Revert "Log info about the caller and callee in hiddenapi denial..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb2b0a1e-355a-4993-9d6b-8701b740e438). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: Ic1d1e80f81d7d8c680759af8920a75464c5892c2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com