Revert "Use libarttest(d)_external to fix 674-hiddenapi, and ena..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb2b0a1e-355a-4993-9d6b-8701b740e438). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: I4b749ae31aeca899f84a5853d20c7f227b8d22fb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com