Revert "Log info about the caller and callee in hiddenapi denial..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9c28209-50d0-46f7-bf89-1f874a8c9316). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: Id4b6a3c9e83c3360db087b435a2393fecf5f41d0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com