Revert "Fix the loading of libarttest(d).so in 150-loadlibrary a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9c28209-50d0-46f7-bf89-1f874a8c9316). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: I86b5211ea56fb0563e6817dc8272e9eaae6f1329
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com