Revert "Use libarttest(d)_external to fix 656-annotation-lookup-..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9c28209-50d0-46f7-bf89-1f874a8c9316). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: Ia60898dfb38d2883bd875b4bf0ed86be65054c5a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com