Revert "Use libarttest(d)_external to fix 674-hiddenapi, and ena..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9c28209-50d0-46f7-bf89-1f874a8c9316). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395168644
Change-Id: Idd7e940119322098cffaba17787dd6b2b200378b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com