Revert "Optimize FindVisitedBlockWithRecyclableSet"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f3f28663-746d-4cc8-927f-53bb9fb47c52). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4b28c6f3444187e5dcc93d9d69122c948c8f570f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com