Revert "Extend suspend timeout for debug activities"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cde84e00-add5-4e90-943f-16a3e7343bee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I13b53d083776a312642e169d29bc6fb0d9393e8c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com