Revert "Extend suspend timeout for debug activities"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8860029-62c4-416b-98bc-24c84e159756). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic0ff64d1c47820521d02f4aa06e2a02de3f8a984
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com