Revert "Refactor ElfFile."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3a9c2e50-5c43-4d8e-9ae7-ae3ac06d6b8e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic7195cf2c8d573445e95e8bdc97895942a4b9f31
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com