Revert "Optimize HConstantFoldingVisitor::PropagateValue"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3a9c2e50-5c43-4d8e-9ae7-ae3ac06d6b8e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I600421c55213b8149f4270db74d9db0ce2c87fd1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com