Revert "Hide most symbols in `ClassLinker`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b6a939ea-ea24-415b-b056-33d6a22b8dae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 397157422
Change-Id: I84af21ecdce6c3fab2995625f904b1d998825b46
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com