Revert "verifier: Reject `filled-new-array/-range` with `[J`/`[D`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/92300175-0951-45f7-877e-026cffe33d48). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieaf89b3b1bbab1999fb54a8f9a08825f2f37933a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com