Revert "Avoids copy using float registers on aarch64 for faster ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/63a887e5-9e00-460b-8ee7-e6263e4e34c8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie4497feb255b39fde09d16e57715e88f442ef2dc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com