Revert "Fix OatFileAssistant non-determinism in choosing best oa..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f924e43e-823c-4687-878d-9ffc5eba3155). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1319a5bb1e1aa4cabe60bf6e244399c850a6fd19
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com