Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0f5feeec-84df-4636-b6fe-cb51ba7b8a0c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e969b6a4eb90be557cd8a8dc8915bc007e2e52b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com