Revert "Extend suspend timeout for debug activities"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bfce784e-d3c4-4646-ab2b-dadb4ca6761d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic8fbefd62166f34db69637c377eea15fe115235e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com