Revert "Extend suspend timeout for debug activities"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5700c57e-43df-4d22-b850-5fd6b3f38067). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9ac8022b82534dbbe450e330b56c830b1f4fa53e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com