Revert "Optimizing: Speed up `HInstruction::Add{,Env}UseAt()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e3758a8c-5eb1-4c4a-845d-82f42a2c59b8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I09278726027928cbeda325715d3deabb53518094
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com