Revert "Optimizing: Speed up `HInstruction::Add{,Env}UseAt()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8573e18d-107a-4943-b1b6-d067bf1c4ebc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idaac506fc2ac886f2d18397818f8ce1bb75e38ff
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com