Revert "Temporarily enable post-gc heap verification"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d6b15860-84f3-4053-a2b1-a0f82f172651). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 397818142
Change-Id: I3d46f7e020f747b991c94cd5d1e4685028f572cf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com