Revert "Extend suspend timeout for debug activities"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5240922f-8172-48ca-bb0c-d988cd90ba65). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8ac20c7da1a543c6cc1dc635559eab72dbdf3bfc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com