Revert "Temporarily enable post-gc heap verification"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0c849728-356e-4568-be08-e17bc9ff4cea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 397853036
Change-Id: I64413249071ff530e23ac008ef0cd96b76901d5e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com