Revert "Fix lock level violations when dumping low overhead traces"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fb30f129-3ba4-4548-a052-f9c8b5cf7705). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic41036b10587daf8b3eab59f9e8dbad357a1971d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com