Revert "dex2oat: Change reported data for `--dump-timings`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/94c5752c-e438-4dbc-8e26-ecfbeaf6dae7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a846079c8546850fe7b1ffe2d951f73e71d4904
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com