Revert "dex2oat: Change reported data for `--dump-timings`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/03c1d7c7-743d-40c5-b47e-e41d4a098df6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I647cbc5ea93302a54654e5993ee068636c03f9a3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com