Revert "Fix OatFileAssistant non-determinism in choosing best oa..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/84993414-268a-40a5-9efc-686a41f8f2e5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib1792c8f9508539f5e66f61df841408cfb5b5045
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com