Revert "Convert cpplint-art-all to Android.bp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed2a73c5-9d59-4ce5-833c-18a02abd1e6d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 398985398
Change-Id: If2b1a47a9d22dadd9185e29cd7b1b19072da3c50
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com