Revert "Convert cpplint-art-all to Android.bp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b64d7f0b-8d58-4563-97d6-0ec1b8405358). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 398984904
Change-Id: Ib0fbe039e12df3356377e891c13e97dc593195df
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com