Revert "Fix lock level violations when dumping low overhead traces"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3040dadc-b68f-4eb4-a26e-2bc88542d28e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib04daef9a34f7465982cf2126947e162dbb4e32f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com