Revert "Convert cpplint-art-all to Android.bp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c639bed7-7261-490d-859c-a86b834760ee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I479572e63af456cad5159a54adbe1a1cb02722a1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com