Revert "Enable pre-GC heap-verification for debug builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c27d87af-fca8-4c58-af85-ebef1cfe78c2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 397044440
Change-Id: If6a415140b21ddf675de09000eb08f5f916013bb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com