Revert "Avoid implicit conversion to bool in SideEffect's methods"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f5d91aee-7b6d-44fe-bfa9-5a0799f80f8d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7703fe6c25b9e689b89cc6728be370f035564c86
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com