Revert "Add card-table verification at the end of GC for generat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b5cdd3d-e64b-4889-8db1-96617dfade13). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I117aaee76f8694b56becb66feffcdd0e7817f95c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com