Revert "Add card-table verification at the end of GC for generat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d0782d58-382f-44b6-8cd8-b2e4ba15fef8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8f41bf377ede676287a361de40c1905136e9e4f3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com