Revert "Add card-table verification at the end of GC for generat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4114aa15-0cc6-4776-b4ff-a6fa902ae5d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I72267ca1157928912ce35aa375a9f8168274b031
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com