Revert "Extend suspend timeout for debug activities"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c9b3001e-113e-445a-b6bd-dc32a709882a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idccacf87c2f52aae7a81d3e0836f77bb44fafd9c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com