Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a8d2971-f29d-4134-83c7-afea44b2a8fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55f428301180f75a6769b3da35c5df51655ad6d7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com