Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fb52fce8-bf10-4946-8c16-74e3c6ca1159). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4a01282a65f9abcab7a2d3b36975c87a1e74f32b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com