Revert "Enforce core platform hiddenapi checks for Baklava."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83e3d800-3d4d-481e-8e7d-ccf12f5bb49a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2ece0d38f1f4668f0849bb97a3aacb502c1945cc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com