Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9ca62f8-2206-4738-8a51-421846d143b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic55e418b5ef31aabf20aae76b3f76572c3160d7b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com