Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8380be12-2ba6-4db8-9463-17fedea0ea56). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia4ef074993eba0c1daaf11c547e8956c6cf3681d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com