Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ab8c49d-ac28-42f4-9a05-43a01cd6b92e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I19a9e0c7500ca6816fe270c0541a27eb02a35e55
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com