Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b972b135-bbf3-4e3e-96d3-75ab2a487159). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e2e40f737f3ec1ccfd22373f3256deb5977acee
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com