Revert "Implement if instructions in fast baseline compiler."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/54241602-fb7e-4602-a435-9cec49ec0c9c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1a3949cf06be0e910eea49e2de8b597955554d5b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com