Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d6cda7ea-de57-4c25-9616-1d976f059553). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic16b1981f9cb1717e09eea049bfffcd6ecb48f8b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com