Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/571662a7-7bad-438e-a153-3fd272ceb24d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f1718d5a5e366cbca83e3a46fc6e58cfe49112f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com