Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/972e3645-48dc-43a3-b0d9-b49c7943e81a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I31f04e7d05af18a5b8f2b58a6fcd0ae03d7eaba4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com