Revert "Turn string raw in testrunner.py"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f3eeacef-231d-4716-93a1-96d768736640). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I91338d8abbf3742e55f8d6648a26c2db6a76fa27
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com