Revert "Fast field lookup in nterp."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f0cb26c5-6575-4e36-a437-a2e893e0460b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifd5d316a70640c1f1e6072b89e94deed1e5dc67d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com