Revert "Fast field lookup in nterp."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6e83d1b-7939-4b2d-b7ef-ea5f4401f716). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id189c7c275801520d94cf48f2fca957467b5af22
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com