Fix the logic of adjusting class-after-obj map in FinishPhase()
It was being wrongly assumed that while iterating over the map, once an
object (which is the value in the map) is found to be >= mid-gen-end,
then we can just erase all entries afterwards. However, that's
incorrect as we could have a next entry in the map with its object <
mid-gen-end, and hence shall be retained.

Made some other corrections also regarding the case when compaction is
not performed.

Bug: 399874996
Bug: 343220989
Bug: 397381582
Test: ATP test-infra
Change-Id: I73e872bfad97a2dc99d2eee44f06c3ce7868bc5c