Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2df16911-c9d1-432a-ac50-396515817e18). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I12f9dcee9bf3f808175add6bc46a592b728dcb21
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com