Revert "Turn string raw in testrunner.py"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aa2b470c-f771-4cd4-a241-aa0244c2e682). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic25d41d15b7f77cb3836b36e31b49ef592121b05
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com