Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4be7dfcc-05f6-41f7-bab3-8bc81221a3f5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f666afdb8fc7a8868c311b585514be09fbe660b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com