Revert "Implement if instructions in fast baseline compiler."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6c222219-bf53-44e3-b7dd-a107886b5b80). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia6cc390de7ebea0130ffdac24535a26ce4f1c851
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com