Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e4e16de9-d137-4fa4-980f-84853beca6f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1927c9df5e36701040a447eb4c31cff534bd671d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com