Revert "Compute references for unreachable instances too."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/900985b4-85d3-4e19-af78-85a39f4d61ba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I67ec40a280b6cf6b96436239009ddb7bc73fa081
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com