Revert "Optimizing: Speed up SSA Liveness Analysis."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8680c2ef-95b0-4e75-baf2-723c475f71d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I33fdb071ce282113347c0d2d3f42577b5a00a5ef
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com