Revert "Age dirty cards in MarkingPause for full GCs as well"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bd879dc4-79e3-468e-a605-2887c24b5cfd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d9ecc279be0b1c0cf3332f32c052702f52c43e6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com