Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5dec6678-d53b-4a48-a362-70e672bd313d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b86613557d26728b2eb970bf2aca4b8ecc2b625
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com