Revert "Optimizing: Speed up SSA Liveness Analysis."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1fe596b3-07e3-4565-b291-c2e5b2052849). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id72738adf2b825b9e5228a68aad099091d16da8b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com