Revert "Age dirty cards in MarkingPause for full GCs as well"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c141792a-b41d-4abc-922b-896a74e5cb3f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5ab7cf63845ff3a0db4cf91a1d70f1df78068310
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com