Exclude non-deterministic fields from oat header key value store.
For Cloud Compilation, we need to enable hosts to generate a boot image
that has exactly the same checksum as generated on device. In other
words, we need to make the oat checksum deterministic across hosts and
devices.

To achieve that, we need to exclude the fields in the oat header key
value store that may differ.

Bug: 372868052
Test: m test-art-host-gtest
Test: Generate a boot image on host. See the checksum being identical to
  the one on device.
Change-Id: I9f73a28b349b673c25909b61a0c7ae8cf883c014