Ensure oat checksum determinism across hosts and devices.
For Cloud Compilation, we need to enable hosts to generate a boot image
that has exactly the same checksum as generated on device. In other
words, we need to make the oat checksum deterministic across hosts and
devices.

To achieve that, when writing the oat header, the non-deterministic
fields are padded to their length limits and excluded from the oat
checksum computation.

Bug: 372868052
Test: m test-art-host-gtest
Test: Generate a boot image on host. See the checksum being identical to
  the one on device.
Change-Id: I9f73a28b349b673c25909b61a0c7ae8cf883c014