Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81c275d6-d7af-447c-a9d4-c8728b640edc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I728e991beec08bc65de4b81db40ecc35ea19473a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com