Revert "Improve reference verification at end of GC cycle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89decc22-bac6-40ec-b43f-a7a90c2f2062). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie48c5ee23b40e1f5a1f114162a86e860a18fe61a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com