Pad the oat header to a fixed size.
For Cloud Compilation, we need to enable hosts to generate a boot image
that has exactly the same checksum as generated on device. In other
words, we need to make the oat checksum deterministic across hosts and
devices.

Because the key value store in the oat header is variable length,
depending on the length of the key value store, all the section offsets
in the oat file may differ, causing the checksum to differ.

To solve this problem, this CL pads the oat header to a fixed size.

Bug: 372868052
Test: m test-art-host-gtest
Change-Id: I45ad293f85c975e3cc4b9ae679ef57ab1cadcdf6