Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a01f9db6-701d-4a37-a11e-31fcd80d4042). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4246429d1dae76454be23dad069931bb01cb3ac8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com