Revert "Remove extraneous apex_available from art tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/98d1c406-02b2-4862-8802-d3e9bde71dde). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia5824a997131fb0b409e527ea53ea740378343d5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com