Revert "Improve reference verification at end of GC cycle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8b72ee0c-39de-46fc-b568-843be7b67e40). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia393966d8506a373422b63cea8c353d14993c1ef
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com