Ensure for any object in old-gen its class is also in there
Like in black-dense case, during young collections also we need to make
sure that for any object in old-gen, its class is also in old-gen.
Otheriwse, we run the risk that if some first-object of a page in
old-gen has its class pointer updated, then we'd be reading wrong data
from the class object.

Bug: 401488879
Test: ATP test infra
Change-Id: I518dd732ab3e2ccb91bf5b5578474ebf90665f81