Revert "Enforce core platform hiddenapi checks for Baklava (reland)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21c6dd03-a942-43ec-a884-2d4bfee80992). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 401706425
Change-Id: I71bd7720652f60864c9ce878478786a6ccb7aa06
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com