Revert "Remove extraneous apex_available from art tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b768e824-1bfe-4c0e-a147-ead6ca96fea7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6053c96bfcca4e82253ee4bfa39f706a1a3bc968
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com