Revert "Improve reference verification at end of GC cycle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/98ced452-88b0-4a3b-be6f-5832c867134c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8682ee09966f6795b0fbed85da03dc289070daa0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com