Revert "Improve reference verification at end of GC cycle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0fb29f6c-b769-4b84-9f15-d72ee925e7b9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I029c03c7e6f803af24a69274a7312e6553188f45
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com