Revert "Improve reference verification at end of GC cycle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ad9a522b-f388-4e06-868a-4f7afada31ef). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id5dd4642ec47d05b7e65236fe3d9ca1eb5809dfe
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com