Revert "Improve reference verification at end of GC cycle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d326f946-90fe-4769-8c6b-495180bb2821). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I13e02d6433c0c98c7e432703a85d1323c037c702
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com