Revert "Improve reference verification at end of GC cycle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9b0a3b42-0276-4bf1-9957-0a149bf844ab). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia6b80070ab85426168097bf43ed1ac24a6aff1fb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com