Revert "Improve reference verification at end of GC cycle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/59b82943-ade5-4fd9-be22-4c4ee94513f9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4b816fb6282fff01d206ae141fbafa0ff6d8b4a1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com