Revert "Improve reference verification at end of GC cycle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7c2d0cf3-4b7d-467b-a348-6b03866097a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iade756ec1f969a1c5545c6baa443bddebcb44677
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com