Revert "Verify surviving objects post-GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/914d8010-c429-45b6-a902-b7264a3a5aeb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I91bb4f7fc8b1bbffaa690416bf302304e847f577
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com