Revert "Improve reference verification at end of GC cycle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3d25fc98-3598-430b-9fba-2af33e53d74e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icfde40d98601e95c5b48397650e71875d450efac
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com