Revert "Remove extraneous apex_available from art tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d52ad154-1288-419c-b8b4-1cee512a0996). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I11635da1da9a4332885a94b8d984ddbfa2c20379
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com