Revert "Dump the event time in nanoseconds instead of timestamps"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8df6daba-e4fc-4f90-a432-073d10a0b1ae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I79202def31a35e46a740aecea98a49d455e818f1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com