Revert "Fast field lookup in nterp."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d436f5d-3314-4892-9973-3eef328a681c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4741e86e48109c705b9e54c53e543385976a9c91
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com