Revert "Improve reference verification at end of GC cycle"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a70b6853-6de0-4d90-b798-93a1ad8cffe9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If0ae2e1358a95bd73e1ead5f0142119dac615023
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com