Use sigwaitinfo instead of sigwait when waiting for signals
sigwaitinfo also returns the siginfo_t that allows us to pass
additional information along with the signal number. Currently the
additional information is unused. This will be used when starting
long running method traces for ANR reports. The additional argument
will be used to specify the duration of the trace.

Bug: 352518093
Test: art/test.py
Change-Id: I19fc84b273a17a9b573ee41c545b4d4679ea9b39