Revert "Pass the dex register array to NterpGetInstanceField."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b7b63046-ac08-4372-9816-8e28e3903dd0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 402253994
Change-Id: I30d002cacb4e426aab94b3b28350c73dc09758f0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com