Revert "Fast field lookup in nterp."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1a64006a-7b27-4d80-8d46-17d669e40c55). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I986c21df5aed322d10e5168e613bffc231d04f79
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com