Revert "Fast field lookup in nterp."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/54a3eb06-39b4-4ef5-ae01-133d138db5f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8f48975f4571be11067a1650d1cfc640c7d27a7f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com