Revert "Remove SDM status from dump."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bd4827c5-1386-4a98-9512-674f834cfcca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic429a0b56f0813699364eafd332d5a195dbd2422
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com