Revert "Ensure oat checksum determinism across hosts and devices."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bd4827c5-1386-4a98-9512-674f834cfcca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie269ec64ecefc46f9e8894c0de23bb134137d9d0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com