Revert "Remove more dead code for ELF file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5b798c07-1880-4943-bdc6-a17c325a1469). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia422ee1ed48248272b9d99ce69bd465d6142699d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com