Revert "Remove dead code for ELF file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5b798c07-1880-4943-bdc6-a17c325a1469). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61078249cc61b0a10cb5b3171142cf17fe18080b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com