Revert "Disable post-GC verifications for non-debug builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e59ed25e-d8a8-4739-af62-041fc3d50ab3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2333049c3e7bb0ef2ee67b1346a088ff29d1a591
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com