Revert "Disable post-GC verifications for non-debug builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bf366964-6d8b-4571-ab3b-98b57746292b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I173ea3275569b50d8ef2265315cae8b0241bdf31
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com