Revert "Compute references for unreachable instances too."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd9869ff-bd6d-4ed1-988c-753b22e4ba17). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I40933484a3081088a304ac169df2868f37885696
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com