Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d929282b-0798-4322-a5d2-d896d275e245). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id61226d5c8347108792077ba2da3234f44cc0173
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com