Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4cfac99a-6190-4d59-8235-af2b5aa6dd93). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I75d2808fdb553b87f052bb24bfbedc92f92b4088
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com