Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6372e0a7-fed2-4052-bb31-9526009361fc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I591be83b1e1469668e91cf1fc12c9c1f4d202d55
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com