Revert "Disable post-GC verifications for non-debug builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7998ba0d-317d-4a5b-8d13-1fd2437ec2ad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib2d23dda26559997110bd6204ea3091d2e47d502
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com