Revert "Disable post-GC verifications for non-debug builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/764113dd-1a9a-4b18-b8e9-f2e5fc9392a8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0e7be286d4615945a93816b7094da7673da4814
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com