Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7946540c-a4a6-4430-b80d-40ed575d996d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieeaa687e7117e5655ce74b8fa5c60ee1622d4d19
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com