Revert "1963-add-to-dex-classloader-in-memory: remove workaround."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d71ce4d6-ad44-48b4-a8a9-b7cc5ff55ac7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie3bd37336cbea9cc774af8ffe28652c1fe5f6b08
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com