Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e94b5936-846c-48d2-895a-3a4f59cea33a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib86090e9b1c9533e8f222174ab3a21836a6bc8a8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com