Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5ddc6a9d-a4a3-4722-8413-f27885e39d73). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I08a8fc9dc2a375f8ecb2803d5452adc16b444cb6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com