Revert "Clear thread-local interpreter cache, instead of sweeping it"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2169fdeb-eb27-42e9-9f5f-172a41b24716). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If5dfbba4e2226db3ec3851ef37bdc07339c4e014
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com