Revert "Disable post-GC verifications for non-debug builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/504c23f4-87b6-4c55-9fe1-20f1efcc1e69). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iff05514f64741440fb265b8594ed14c01c41f7c4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com