Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7483f4f4-b5a8-4d46-b968-ab9e8ac445e3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2e438fc165adf9578b1bcd3de3eb9f800a22c515
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com