Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3b61881d-27f4-475f-be32-ccd05f1cadfc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia36e87c1dfa6deb95f9449f596b92e7ddd2bb43c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com