Revert "Disable 055-enum-performance on tradefed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/101929aa-88f4-4eeb-b53a-63ed3dcfc439). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7ab466ed3081a7474a98809b97123df5b7e09790
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com