Revert "Disable post-GC verifications for non-debug builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9bc95bc2-90aa-4868-aed9-e6e8406f1760). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I76e804908b44efea8861a5cedc992ccda93732e9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com