Revert "Disable 055-enum-performance on tradefed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a7c82283-0e00-4370-a7be-d97b8f8763d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ief6c7e19730b807bef6eb7f7546bbaae5392c9a2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com