Add _BIONIC_ASSERT_MALLOC_USABLE_SIZE_IS_SAFE to allow _FORTIFY_SOURCE=3 with malloc_usable_size
We plan to make _FORTIFY_SOURCE=3 global on device and ti_allocator.cc
in art/openjdkjvmti uses malloc_usable_size. This specific use of
malloc_usable_size is allowed with _FORTIFY_SOURCE=3 because it is only
keeping track of the allocation/deallocation byte counts. As a result,
we define _BIONIC_ASSERT_MALLOC_USABLE_SIZE_IS_SAFE to avoid the clang
error.

Bug: 291762537
Test: m with aosp/3539100
Change-Id: Ic730efd21fc54d1dd0a584a06ab8c11b7678cc90