Revert "Add app/GC interference metrics in ART"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/95670f09-f6e3-4570-b8c3-b49285ba0e66). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib778316e0571c70ccc51c8e0df252875e5f543d9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com