Revert "Use right type when using unique_ptr for array"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/95670f09-f6e3-4570-b8c3-b49285ba0e66). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3137657de52b5220d5094a24b487c48623620724
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com