Revert "Revert^2 "Call target method in accessor MHs when it is ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/989e20f3-99f1-4d8e-8a12-0ca7eb8dd388). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 405326442
Change-Id: Iaca99587255dd4fc549ad67b9ec1288796640ece
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com